fix: restrict code argument
